chore: team ui setup
